+++ ft/commands.go
@@ -7,7 +7,6 @@ import (
 	"net/http"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"runtime"
 	"strings"
 
@@ -78,90 +77,6 @@ func PassCmd(args []string) (*Cmd, error) {
 	}
 
 	return cmd, nil
-}
-
-/*
-Takes a key or relative path and returns it's absolute path or an error.
-
-Path is a seperate argument to handle evaluating many paths as part of a loop.
-*/
-func evalPath(data *CmdAPI, path *string) (string, error) {
-
-	// TODO
-	//  - check for SPECIFIC errors in directory checks (ex: *PathError)
-	//  - ensure other types of errors are caught and returned early
-
-	var key string
-	provided_string := *path
-
-	if strings.Contains(provided_string, "/") {
-
-		path_array := strings.Split(provided_string, "/")
-		eval_array := make([]string, len(path_array))
-
-		// key evaluation if the first string before "/" delimeter is key
-		key = path_array[0]
-		if p, ok := data.allPaths[key]; ok {
-			eval_array[0] = p
-		} else {
-			eval_array[0] = key
-		}
-		for i, str := range path_array {
-			if i != 0 {
-				eval_array[i] = str
-			}
-		}
-
-		// handles evaluated path and relative paths
-		path := strings.Join(eval_array, "/")
-		dir, err := os.Stat(path)
-		if err == nil {
-			if dir.IsDir() {
-				return fmt.Sprintf("%s", path), nil
-			}
-		}
-
-		return "", errors.New(fmt.Sprintf(InvalidDirectoryMsg, provided_string, path))
-
-	} else {
-
-		key = provided_string
-		// handles key lookup
-		p, ok := data.allPaths[key]
-		if !ok {
-			// handles releative directory in CWD
-			dir, err := os.Stat(key)
-			if err == nil {
-				if dir.IsDir() {
-					// in this case key is assumed to be a relative path
-					p, err = filepath.Abs(key)
-					if err != nil {
-						return "", err
-					}
-					return fmt.Sprintf("%s", p), nil
-				}
-			}
-
-			// handles CDPATH
-			cdpath := os.Getenv("CDPATH")
-			if cdpath != "" {
-				cdpaths := strings.Split(cdpath, ":")
-				for _, path := range cdpaths {
-					cdPathResult := filepath.Join(path, key)
-					dir, err := os.Stat(cdPathResult)
-					if err == nil && dir.IsDir() {
-						return fmt.Sprintf("%s", cdPathResult), nil
-					}
-				}
-			}
-
-			return "", errors.New(fmt.Sprintf(UnrecognizedKeyMsg, key))
-
-		}
-
-		return fmt.Sprintf("%s", p), nil
-
-	}
 }
 
 // changeDirectory can handle key lookup, relative paths, directories in CDPATH, and key evaluation.
