+++ ft/settings.go
@@ -1,0 +1,8 @@
+package ft
+
+// TODO create settings data structures
+
+// Settings
+
+// - toggle for history across active sessions (think tmux workflows)
+// - auto updates (or check for updates)
